+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - (Debug Package) Initial commit
 - (Feature) Detach PVC from deployment in Ordered indexing method
 - (Feature) OPS Alerts
+- (Feature) ScaleDown Candidate
 
 ## [1.2.16](https://github.com/arangodb/kube-arangodb/tree/1.2.16) (2022-09-14)
 - (Feature) Add ArangoDeployment ServerGroupStatus